created table component & table-service
