+++ clinic/src/app/core/auth/table-data-service.service.ts
@@ -1,0 +1,31 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class TableDataService {
+  tableData: { cols: { value: string; activated: boolean }[] }[] = [];
+  tableHeaders: { num: number; day: string }[] = [
+    { num: 17, day: 'mon' },
+    { num: 18, day: 'tue' },
+    { num: 19, day: 'wed' },
+    { num: 20, day: 'thu' },
+    { num: 21, day: 'fri' },
+    { num: 22, day: 'sat' },
+    { num: 23, day: 'sun' },
+  ];
+
+  constructor() {
+    this.initializeTableData();
+  }
+
+  private initializeTableData() {
+    for (let i = 1; i <= 9; i++) {
+      const row = { cols: [] as { value: string; activated: boolean }[] };
+      for (let j = 1; j <= 7; j++) {
+        row.cols.push({ value: `${i}-${j}`, activated: false });
+      }
+      this.tableData.push(row);
+    }
+  }
+}
