+++ clinic/src/app/custom-table/custom-table.component.ts
@@ -1,0 +1,24 @@
+import { Component, Input } from '@angular/core';
+
+@Component({
+  selector: 'app-custom-table',
+  templateUrl: './custom-table.component.html',
+  styleUrls: ['./custom-table.component.css']
+})
+export class CustomTableComponent {
+  @Input() tableData: { cols: { value: string; activated: boolean }[] }[] = [];
+  @Input() tableHeaders: { num: number; day: string }[] = [];
+
+  constructor() { }
+
+  getTimeRange(rowNumber: number): string {
+    const startTime = 9;
+    const endTime = 18;
+    const timeSlot = 1;
+
+    const startHour = startTime + rowNumber * timeSlot;
+    const endHour = startHour + timeSlot;
+
+    return `${startHour}:00 - ${endHour}:00`;
+  }
+}
