+++ clinic/src/app/doctor-page/doctor-page.component.html
@@ -67,46 +67,8 @@
         </div>
       </div>
       <!-- calendar -->
-      <div class="custom-table">
-        <table>
-          <thead class="header">
-            <tr>
-              <th class="days-botton">
-                <span class="img-2"
-                  ><img
-                    src="../../assets/right-arrow_271228.png"
-                    style="width: 12px; height: 12px"
-                    alt=""
-                /></span>
-                <span
-                  ><img
-                    src="../../assets/right-arrow_271228.png"
-                    style="width: 12px; height: 12px"
-                    alt=""
-                /></span>
-              </th>
-              <th *ngFor="let num of tableHeaders; let i = index">
-                <div class="days">
-                  <p class="day-number">{{ num.num }}</p>
-                  <p>({{ num.day }})</p>
-                </div>
-              </th>
-            </tr>
-          </thead>
-          <tbody>
-            <ng-container *ngFor="let row of tableData; let i = index">
-              <tr>
-                <td class="rowhead">{{ getTimeRange(i) }}</td>
-                <!-- Display time range in the first column -->
-                <ng-container *ngFor="let col of row.cols; let j = index">
-                  <td class="tdclick" id="{{'td_' + i + '_' + j}}">
-                  </td>
-                </ng-container>
-              </tr>
-            </ng-container>
-          </tbody>
-        </table>
-      </div>
+      <app-custom-table [tableData]="tableDataService.tableData" [tableHeaders]="tableDataService.tableHeaders"></app-custom-table>
+
       <div class="change-sect">
         <div class="redacted-button">
           <button class="redact">
